Extract TrophyController from UsersController
